:art: FORMAT: rename "types" folder to "models"
- :art: FORMAT: rename interface from prefix "I" to suffix "props"
