+++ src/models/Asset.ts
