+++ src/pages/Project/Row/ProjectRow.tsx
@@ -1,18 +1,18 @@
 import React from "react";
 // model
-import IProjectRow, { ProjectRowType } from "./types/IProjectRow";
-import IProjectRowText from "./types/IProjectRowText";
+import ProjectRowProps, { ProjectRowType } from "./models/ProjectRowProps";
+import ProjectRowTextProps from "./models/ProjectRowTextProps";
 // views
 import ProjectRowText from "./ProjectRowText";
 // Styles
 import "./style.css";
 
-const ProjectRow: React.FC<IProjectRow> = (props: IProjectRow) => {
+const ProjectRow: React.FC<ProjectRowProps> = (props: ProjectRowProps) => {
   let row = <div className="project-row-default" />;
 
   switch (props.type) {
     case ProjectRowType.Text:
-      row = ProjectRowText(props as IProjectRowText);
+      row = ProjectRowText(props as ProjectRowTextProps);
       break;
     default:
       break;
