+++ src/pages/Project/Row/ProjectRowText.tsx
@@ -1,7 +1,7 @@
 import React, { ReactElement } from "react";
-import IProjectRowText from "./types/IProjectRowText";
+import ProjectRowTextProps from "./models/ProjectRowTextProps";
 
-const ProjectRowText = (props: IProjectRowText): ReactElement => {
+const ProjectRowText = (props: ProjectRowTextProps): ReactElement => {
   const { content } = props;
   return (
     <div className="project-row-text">
