+++ src/pages/Project/Row/models/ProjectRowTextProps.ts
@@ -1,0 +1,16 @@
+import ProjectRowProps from "./ProjectRowProps";
+
+enum TextStyle {
+  h1,
+  h2,
+  h3,
+  p,
+}
+
+interface ProjectRowTextProps extends ProjectRowProps {
+  content: string;
+  style: TextStyle;
+}
+
+export default ProjectRowTextProps;
+export { TextStyle };
